Move helper functions to trytls.utils
